Switched to double precision for frequencies and periods.
While I was looking for an optimization on estimatePeriod, I found that quite
important errors were made using single precision for autocorrelation.
Therefore I decided to switch to double precision for this purpose, whereas
keep single precision for data acquisition.

+++ include/guitar.h
@@ -94,7 +94,7 @@ extern semitone_t noteToSemitones(const char *name, semitone_t octave);
  *  function
  * @return The semitones of the note from A0
  */
-extern semitone_t frequencyToSemitones(float frequency, float *error);
+extern semitone_t frequencyToSemitones(double frequency, double *error);
 
 /**
  * @brief Get all the frets in which a note can be played in a guitar.
